+++ src/os2datascanner/projects/admin/adminapp/views/exchangescanner_views.py
@@ -13,6 +13,7 @@
 #
 
 from django import forms
+from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 from django.views import View
 from rest_framework.generics import ListAPIView
@@ -81,6 +82,8 @@ class ExchangeScannerCreate(ExchangeScannerBase, ScannerCreate):
     fields = ['name', 'mail_domain', 'schedule', 'exclusion_rule', 'do_ocr',
               'do_last_modified_check', 'rule', 'userlist', 'only_notify_superadmin',
               'service_endpoint', 'organization', 'org_unit', 'keep_false_positives']
+    if settings.MSGRAPH_EWS_AUTH:
+        fields.append("grant")
     type = 'exchange'
 
     def get_success_url(self):
@@ -110,6 +113,8 @@ class ExchangeScannerCopy(ExchangeScannerBase, ScannerCopy):
     fields = ['name', 'mail_domain', 'schedule', 'exclusion_rule', 'do_ocr',
               'do_last_modified_check', 'rule', 'userlist', 'only_notify_superadmin',
               'service_endpoint', 'organization', 'org_unit', 'keep_false_positives']
+    if settings.MSGRAPH_EWS_AUTH:
+        fields.append("grant")
     type = 'exchange'
 
     def get_form(self, form_class=None):
@@ -142,6 +147,8 @@ class ExchangeScannerUpdate(ExchangeScannerBase, ScannerUpdate):
     fields = ['name', 'mail_domain', 'schedule', 'exclusion_rule', 'do_ocr',
               'do_last_modified_check', 'rule', 'userlist', 'only_notify_superadmin',
               'service_endpoint', 'organization', 'org_unit', 'keep_false_positives']
+    if settings.MSGRAPH_EWS_AUTH:
+        fields.append("grant")
     type = 'exchange'
 
     def get_success_url(self):